Test : run k6 separately
