+++ .gitignore
@@ -17,4 +17,4 @@ gradle-app.setting
 .idea/*
 
 **/bin/**
-k6-test/**/results
+k6-test/results
