+++ k6-test/process-cpu-usage-test.js
@@ -4,11 +4,10 @@ import { textSummary } from 'https://jslib.k6.io/k6-summary/0.0.1/index.js';
 
 const uri = "http://localhost:8888/actuator/metrics/process.cpu.usage";
 const expectedMax = 0.7;
-const testName = 'process-cpu-usage-test';
 
 export let options = {
   vus: 1,
-  duration: '1m',
+  duration: '60s',
   thresholds: {
     http_req_failed: ['rate<0.01'], // http errors should be less than 1%
     http_req_duration: ['p(100)<=1000'], // 100% of requests should be maximum of 1000ms
@@ -20,28 +19,10 @@ export default () => {
   const res = http.get(uri);
 
   check(res, {
-    'Status is 200     ': () => {
-      console.info('200----------' + res);
-    },
-  });
-
-  check(res, {
     'Metric value is     ': () => {
-      console.info('metri-------' + res);
-
       const metric = JSON.parse(res.body);
-      if (metric === null) {
-        throw 'metric not found';
-      }
-      if (metric.measurements === null) {
-        throw 'metric.measuarements not found';
-      }
-      if (metric.measurements.length != 1) {
-        throw 'metric.measurement.length is NOT 1';
-      }
-      if (metric.measurements[0] === null) {
-        throw 'metric.measurement[0] is Null';
-      }
+      metricNotNull(metric);
+      console.info('metric -------' + JSON.stringify(metric));
       return metric.measurements[0].value <= expectedMax;
     }
 
@@ -57,19 +38,23 @@ export function handleSummary(data) {
 
   return {
     'stdout': textSummary(data, { indent: ' ', enableColors: true }), // Show the text summary to stdout...
-    'process-cpu-usage-test-summary.json': JSON.stringify(data), // and a JSON with all the details...
+    './results/process-cpu-usage-test-summary.json': JSON.stringify(data), // and a JSON with all the details...
   };
 }
 
-// interface aMetric {
-//   name: string;
-//   description: string;
-//   baseUnit: string,
-//   measurements: aMeasurement[],
-//   availableTags: any[]
-// };
+function metricNotNull(metric, status) {
+  if (metric === null) {
+    throw 'status : ' + status + '  |   metric not found';
+  }
+  if (metric.measurements === null) {
+    throw 'status : ' + status + '  |   metric.measuarements not found';
+  }
+  if (metric.measurements.length != 1) {
+    throw 'status : ' + status + '  |   metric.measurement.length is NOT 1';
+  }
+  if (metric.measurements[0] === null) {
+    throw 'status : ' + status + '  |   metric.measurement[0] is Null';
+  }
+  return metric.measurements[0].value;
+}
 
-// interface aMeasurement {
-//   statistic: string,
-//   value: number
-// }
