+++ k6-test/simple-delivery-get.js
@@ -6,24 +6,23 @@ const uri = "http://localhost:8888/api/delivery";
 
 export let options = {
   vus: 100,
-  gracefulStop: '10s', // do not wait for iterations to finish in the end
   startTime: '10s', // the ramping API test starts a little later
   startRate: 10,
   timeUnit: '1s', // we start at 50 iterations per second
   stages: [
     { duration: '10s', target: 50 },
-    { duration: '10s', target: 300 },
-    { duration: '10s', target: 800 },
+    { duration: '10s', target: 100 },
     { duration: '10s', target: 500 },
+    { duration: '10s', target: 300 },
     { duration: '10s', target: 100 }, // below normal load
   ],
   thresholds: {
     // http errors should be less than 1%
-    http_req_failed: ['rate<=0.01'],
+    http_req_failed: ['rate<=0.05'],
     // 95% of requests should be less than 500ms
-    http_req_duration: ['p(90)<=500', 'p(95)<=750', 'p(100)<=1000'],
+    http_req_duration: ['p(90)<=1000', 'p(95)<=1250', 'p(100)<=1500'],
     // 99% of checkes should pass
-    checks: ['rate>=0.99']
+    checks: ['rate>=0.95']
   }
 };
 
@@ -31,13 +30,11 @@ export default () => {
   const res = http.get(uri);
 
   check(res, {
-    'Status is 200     ': () => {
-      console.info(JSON.parse(res.body));
-      return res.status === 200;
-    },
-
+    'Body length is greater than 1': () => {
+      const array = JSON.parse(res.body);
+      return array.length > 0;
+    }
   });
-
   sleep(0.1);
 };
 
@@ -52,7 +49,7 @@ export function handleSummary(data) {
 
   return {
     'stdout': textSummary(data, { indent: ' ', enableColors: true }), // Show the text summary to stdout...
-    'simple-delivery-get-summary.json': JSON.stringify(data), // and a JSON with all the details...
+    './results/simple-delivery-get-summary.json': JSON.stringify(data), // and a JSON with all the details...
   };
 }
